+++ core-ui/sample/src/main/java/ru/surfstudio/android/core/ui/sample/ui/screen/main/MainPresenter.kt
@@ -2,6 +2,8 @@ package ru.surfstudio.android.core.ui.sample.ui.screen.main
 
 import ru.surfstudio.android.core.mvp.presenter.BasePresenter
 import ru.surfstudio.android.core.mvp.presenter.BasePresenterDependency
+import ru.surfstudio.android.core.ui.navigation.activity.navigator.ActivityNavigator
+import ru.surfstudio.android.core.ui.sample.ui.screen.result.ResultActivityRoute
 import ru.surfstudio.android.dagger.scope.PerScreen
 import javax.inject.Inject
 
@@ -9,13 +11,22 @@ import javax.inject.Inject
  * Презентер главного экрана
  */
 @PerScreen
-internal class MainPresenter @Inject constructor(basePresenterDependency: BasePresenterDependency
+internal class MainPresenter @Inject constructor(
+        basePresenterDependency: BasePresenterDependency,
+        private val activityNavigator: ActivityNavigator
 ) : BasePresenter<MainActivityView>(basePresenterDependency) {
 
     private val sm: MainScreenModel = MainScreenModel()
 
     override fun onLoad(viewRecreated: Boolean) {
         super.onLoad(viewRecreated)
+        subscribe(activityNavigator.observeResult(ResultActivityRoute::class.java)) {
+            view.showToast("Result from ResultActivityRoute is: " + it.isSuccess)
+        }
         view.render(sm)
+    }
+
+    fun startResultActivity() {
+        activityNavigator.startForResult(ResultActivityRoute())
     }
 }