+++ picture-provider/lib-picture-provider/src/main/java/ru/surfstudio/android/picturechooser/Utils.kt
@@ -22,9 +22,12 @@ import android.graphics.Bitmap
 import android.net.Uri
 import android.provider.MediaStore
 import androidx.annotation.WorkerThread
+import io.reactivex.Completable
 import io.reactivex.Observable
+import ru.surfstudio.android.core.ui.navigation.NoScreenDataStub
 import ru.surfstudio.android.core.ui.navigation.ScreenResult
 import ru.surfstudio.android.core.ui.navigation.activity.navigator.ActivityNavigator
+import ru.surfstudio.android.core.ui.navigation.activity.route.ActivityWithResultNoDataRoute
 import ru.surfstudio.android.core.ui.navigation.activity.route.ActivityWithResultRoute
 import ru.surfstudio.android.picturechooser.exceptions.ActionInterruptedException
 import java.io.*
@@ -92,6 +95,19 @@ internal fun <T : Serializable> observeSingleScreenResult(
 ): Observable<T> {
     return activityNavigator.observeResult<T>(route)
             .flatMap { parseScreenResult(it) }
+}
+
+internal fun observeSingleScreenResultNoData(
+        activityNavigator: ActivityNavigator,
+        route: ActivityWithResultNoDataRoute
+): Completable {
+    return activityNavigator.observeResult<NoScreenDataStub>(route)
+            .map {
+                if (!it.isSuccess) {
+                    throw ActionInterruptedException()
+                }
+                true
+            }.ignoreElements()
 }
 
 internal fun <T : Serializable> observeMultipleScreenResult(